Fix test

+++ test/test.cpp
@@ -7,7 +7,7 @@ TEST_CASE ("Text has special characters or numbers between whitespaces")
     string expected = "HelloWorld";
     string result;
     vector<string> words = wordsFromFile("test/1.txt");
-    for (int i = 0; i < words.size(); i++) {
+    for (size_t i = 0; i < words.size(); i++) {
         result += words[i];
     }
     REQUIRE(expected == result);
@@ -18,7 +18,7 @@ TEST_CASE ("Text has special characters or numbers between lettets")
     string expected = "HelloWorld";
     string result;
     vector<string> words = wordsFromFile("test/2.txt");
-    for (int i = 0; i < words.size(); i++) {
+    for (size_t i = 0; i < words.size(); i++) {
         result += words[i];
     }
     REQUIRE(expected == result);
@@ -30,7 +30,7 @@ TEST_CASE ("Words have same beginning and diffetent length")
     string result;
     vector<string> words = {"Alexeyev", "Alexeyevsk", "Alexey"};
     vector<string> sorted = sort(words);
-    for (int i = 0; i < sorted.size(); i++) {
+    for (size_t i = 0; i < sorted.size(); i++) {
         result += sorted[i];
     }
     REQUIRE(expected == result);
@@ -42,7 +42,7 @@ TEST_CASE ("Letters have different cases")
     string result;
     vector<string> words = {"PetrU", "dvAdtstat", "trI", "gOda"};
     vector<string> sorted = sort(words);
-    for (int i = 0; i < sorted.size(); i++) {
+    for (size_t i = 0; i < sorted.size(); i++) {
         result += sorted[i];
     }
     REQUIRE(expected == result);
